requirements.txt
psycopg2-binary==2.8.5
Django==3.1
django-modeltranslation
django-rest-swagger
djangorestframework==3.11.0
django-parler
drf-yasg
djangorestframework-simplejwt
Pillow==8.0
django-3-jet
google-api-python-client==1.4.1
python-decouple